producer/requirements.txt
pika==1.3.1
flask==2.0.1
cloudevents
python-dotenv
